Change structure & clean up implementation of commands & sequences
